internal/ci: top of stack commit for refactors
This dummy commit acts as the top of stack commit for refactors of
internal/ci. It is the final commit prior to the refactors required for
the new cmd/cueckoo setup.

Having such a CL allows other repos to reference this change, but the
change also acts as a reference point for the refactors the other repos
need to make prior to the cmd/cueckoo refactor.

DO NOT REVIEW.
DO NOT SUBMIT.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: I59c0b8e4fb2068ff1fa1b43beda130187102a744