internal/ci: collapse gerrithub package into base
The separation has long been rather awkward, and with hindsight was
definition a premature optimisation or attempt at spotting an API before
a pattern emerges.

If at a later date the need for a split is real, an API for a good split
will likely present itself.

This change should not result in any .github/workflows changes.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: I4710facae844216c78b0693d090d82b8c9ce1f9a