internal/ci: collapse gerrithub package into base
The separation has long been rather awkward, and with hindsight was
definition a premature optimisation or attempt at spotting an API before
a pattern emerges.

If at a later date the need for a split is real, an API for a good split
will likely present itself.

This change should not result in any .github/workflows changes.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: I667804b2d32c6ecda3bf83e30b1d4a8e602bff56