internal/ci: split base into separate files
Until we establish a clear requirement for better package boundaries,
base is the one place where we put common definitions (gerrithub is
folded into this package in a later CL).

Split the package into different files to help separate the different
aspects of the package.

This is a purely mechanical change.

This change should not result in any .github/workflows changes.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: Idb8f969a81732eec19f045609bc640d0b62a5be2
Reviewed-on: https://review.gerrithub.io/c/myitcvscratch/cue-gerrit-test/+/551682