internal/ci: move away from using defs as params in base
This is a purely mechanical change to shift from using definitions to
regular fields for the "params" of base.

We embed base into core in order to validate that core satisfies those
parameters.

We finally confirm this change as working by switching to use core
instead of explicit parameters in workflows.cue

This change should not result in any .github/workflows changes.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: Ic7d385b9bd3926177ccd71009b08f5915d3a3086